requirements.txt
Flask==3.0.3
PyYAML
python-dotenv
openai
flask-cors

